requirements.txt
feedparser==5.2.1
python-telegram-bot==8.0
